Enable make clean
Enable the clean target in the Makefile to clean the artifacts generated
during the build and coverage tests.  Add the bin dir to .gitignore.

Before this change the makefile had a clean target that performed
'git clean -dx' in a source directory, which is a dangerous operation in
a development environment since it will discard any uncommitted source
code changes that a developer has made; it also did not clean up the
actual build/test artifacts since they were not generated into the
directoried being cleaned.

Change-Id: I23fd5f44c8ce1cf6f539c0e05ae91c667a34e044