SSO ability for plugin dashboards
This facilitates the SSO solution for plugin dashboards
To accomplish this we must use a reverse proxy to
inject the correct header information on requests
so that the backend services will accept a prelogin.
This is specific to each plugin so it needs to be
extensible and allow for a dashboard to customize how
it's done.

This also allows for a fully hosted page when running
in headless mode.

Fixes #42

Change-Id: I442e22a19a387ae367c2518085cdca0f2786a0a4