Remove electron from the tree
To quote Mike LaVette a man with 2 watches doesn't know what time
it is.  Now that we have the ability to run in the browser the
purpose of electron is that of a stepping stone and as such is
relegated to the antiquity role for this project.  We now have
only 1 watch.

Change-Id: I9f61459b8509d1afbea5ca5949a58542edfcd895