Refactor yaml explorer
Fixes #50

The yaml files exposed in the editor are now based on the phase
configuration specified in the metadataPath key of the airship
config file, and associated phase documents. Changes have been
made to the backend to allow for on-demand loading of phase
bundle documents, which will be necessary for an upcoming
change implementing the rendered document viewer.

Change-Id: I6d1a586e500fb9d82a65f5b73de6ca85bdff824c