Add rendered document viewer to UI
A POC for a rendered document viewer for airshipui.
By default, the viewer shows phase details (if there
are any), and provides access to the executor document,
the fully rendered phase bundle, as well as all individual
documents that make up the bundle. Filtering of individual
documents can be done using Selectors. All documents in
this view are read-only, and will probably stay that way
unless there is a well-defined use case for making them
editable.

Currently, this change is dependent upon the merging of
https://review.opendev.org/#/c/752764/ to fix a bug in
the ListPhases function.

Change-Id: Iab40db2f721adc94aa93a733efc6119fde40a55c