Add additional phase actions
Adds an action menu next to each phase in the tree
view to allow users to view the rendered phase bundle,
validate the phase (not yet implemented in ctl), and
run the phase using run options. Events are handled
by an event processor which displays events as toasts,
but this may change in later iterations. Log messages
appear in the log tray in the footer.

Note: for testing, it may be necessary to modify the
airship config file to remove "airshipctl" from the
targetPath and prepend it to metadataPath and subPath.

Change-Id: Iac4439fc21409bffca1f95a03425b95135fa7d0d